Refreactored RecycleViewHolder added searchable
