+++ app/src/main/java/pizzarat/cs2340/gatech/edu/ratpatrol/SearchArchive.java
@@ -1,0 +1,12 @@
+package pizzarat.cs2340.gatech.edu.ratpatrol;
+
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * Created by evieb on 10/29/2017.
+ */
+
+public class SearchArchive extends AppCompatActivity {
+
+
+}
