+++ app/src/main/res/layout/activity_archive.xml
@@ -10,10 +10,29 @@
                 android:paddingTop="16dp"
                 tools:context="pizzarat.cs2340.gatech.edu.ratpatrol.ArchiveActivity">
 
+    <SearchView
+        android:id="@+id/searchView"
+        android:layout_width="wrap_content"
+        android:layout_height="45dp"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:paddingBottom="10dp"
+        android:visibility="visible">
+
+    </SearchView>
+
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
+        android:layout_below="@+id/searchView"
+        android:clipChildren="false"
+        android:clipToPadding="true"
+        android:paddingTop="0dp"
         android:scrollbars="vertical">
 
     </android.support.v7.widget.RecyclerView>
